fixed environment variables error
Description: 1. when changing .env file variables, the theme was not
changing. 2. when changing UI_DIRECTION key, the view was still
left-to-right.

Error: 1. when reading .env file, the returned object had two props:
parsed and error. Instead of setting the parsed to the env key, I was
setting the entire object. 2. the env key was correct, but the css was
still direction:ltr.

Solutions: 1. in next.config file, setting the `env` key to the parsed
value and not the entire object. 2. adding to the `body` tag in
_document file the `dir` attribute and setting it to `theme.direction`.